Fixed overflow issues for fft

+++ src/c_dsp/fft.h
@@ -11,11 +11,12 @@
 
 #define N_TWIDDLES N_FFT/2
 
+
 void fftTop(DPRAM_64 &dpram0, DPRAM_64 &dpram1);
 
 void butterfly(int32_t inReal1, int32_t inImag1, int32_t inReal2, int32_t inImag2, int32_t twiddleReal, int32_t twiddleImag, int32_t &outReal1, int32_t &outImag1, int32_t &outReal2, int32_t &outImag2);
 
-void complexMultiply(int32_t inReal1, int32_t inImag1, int32_t inReal2, int32_t inImag2, int32_t &outReal, int32_t &outImag);
+void complexMultiply(int32_t inReal1, int32_t inImag1, int32_t inReal2, int32_t inImag2, int64_t &outReal, int64_t &outImag);
 
 // Perform a circular left shift on N bit word x by y bits
 void rotateLeft(uint32_t &x, uint32_t y, uint8_t N);
